Fix minor number in log message
Some operating systems on the Linux kernel use the USB subsystem
with dynamic minor allocation for any devices that uses the USB major
number. If it is not enabled, the minor number will be based on the
next available free minor, starting at the class_driver->minor_base

This commit takes this into account in debug messages.

Signed-off-by: Andrey Zakharenkov <a.zakharenkov@yadro.com>
Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>
Tested-by: Evgeniy Chudaykin <e.chudaykin@yadro.com>
Reviewed-by: Maxim Polyakov <m.polyakov@yadro.com>