Add interface to reset EC USB controller
Commit adds function to reset the USB interface from the host to the
microcontroller. This is necessary when the USB controller on the EC
does not respond  or any error occurred during the data transfer. To
reset the EC from user-space use ioctl with EC_USB_RESET command ID.

Resolved: https://github.com/YADRO-KNS/fabric-ec/issues/1

Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>
Tested-on: stm32f0-discovery